[SM] Update test entrypoint
- entrypoint now uses updated method getLastSample() to get samples from sensors
- deleted copy constructor from sensor manager and sensor sampelrs
- added cast in TestSensor to avoid compiler warning on getTick() usage

+++ src/shared/utils/testutils/TestSensor.h
@@ -34,9 +34,9 @@ struct TestData : public TimestampData
 {
     float value;
 
-    TestData(float v) : TimestampData{getTick()}, value(v) {}
+    TestData(float v) : TimestampData{static_cast<uint64_t>(getTick())}, value(v) {}
 
-    TestData() : TimestampData{getTick()}, value(0.0) {}
+    TestData() : TimestampData{static_cast<uint64_t>(getTick())}, value(0.0) {}
 };
 
 class TestSensor : public Sensor<TestData>
